api/requirements-freeze.txt
attrs==20.3.0
boto3==1.17.7
botocore==1.20.7
click==7.1.2
Flask==1.1.2
flask-expects-json==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==1.1.1
pyrsistent==0.17.3
python-dateutil==2.8.1
s3transfer==0.3.4
six==1.15.0
ulid-py==1.1.0
urllib3==1.26.3
Werkzeug==1.0.1
